Add documentation to main menu?
As a user looking for the documentation, I think that it would be convenient to have a prominent link to the documentation in the main menu at the top of the page.

Currently I have to scroll through a bit of text to find it. If there isn't enough room, I would propose removing 'Archive'? (or at least renaming it to something like 'library' or something that would indicate it contains a list of books).