Lightweight Machine Learning Classics with R
A first draft of my new textbook Lightweight Machine Learning Classics with R is now available at https://lmlcr.gagolewski.com/

Could this be listed at bookdown.org? Thanks!!!

About. Explore some of the most fundamental algorithms which have stood the test of time and provide the basis for innovative solutions in data-driven AI. Learn how to use the R language for implementing various stages of data processing and modelling activities. Appreciate mathematics as the universal language for formalising data-intense problems and communicating their solutions. The book is for you if you're yet to be fluent with university-level linear algebra, calculus and probability theory or you've forgotten all the maths you've ever learned, and are seeking a gentle, yet thorough, introduction to the topic.